pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "als-rag"
version = "0.1.0"
description = "Retrieval-Augmented Generation system for ALS (Amyotrophic Lateral Sclerosis) research literature"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "ALS-RAG Contributors"}
]
keywords = [
    "ALS",
    "amyotrophic-lateral-sclerosis",
    "motor-neuron-disease",
    "RAG",
    "retrieval-augmented-generation",
    "scientific-literature",
    "neuroscience",
    "PubMed",
    "FAISS",
    "NLP",
    "clinical-research"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Healthcare Industry",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "scipy>=1.7.0",
    "faiss-cpu>=1.7.0",
    "transformers>=4.20.0",
    "sentence-transformers>=2.2.0",
    "torch>=1.10.0",
    "openai>=1.0.0",
    "python-dotenv>=0.19.0",
    "requests>=2.26.0",
    "beautifulsoup4>=4.10.0",
    "lxml>=4.6.0",
    "tqdm>=4.62.0",
    "pydantic>=2.0.0",
    "tenacity>=8.0.0",
    "streamlit>=1.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "pytest-mock>=3.6.0",
    "black>=22.0.0",
    "pylint>=2.12.0",
    "mypy>=0.950",
    "isort>=5.10.0",
    "ipython>=8.0.0",
    "jupyter>=1.0.0",
]

knowledge-graph = [
    "neo4j>=5.0.0",
    "redis>=4.0.0",
]

full = [
    "als-rag[dev,knowledge-graph]",
]

[project.urls]
Homepage = "https://github.com/hkevin01/als-rag"
Repository = "https://github.com/hkevin01/als-rag"
Issues = "https://github.com/hkevin01/als-rag/issues"

[project.scripts]
als-rag = "als_rag.cli.query:main"

[tool.setuptools]
packages = ["als_rag"]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
als_rag = ["py.typed"]

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --tb=short"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "benchmark: marks tests as performance benchmarks",
]
